+++ R/Fire.R
@@ -4,66 +4,101 @@ NULL
 
 #' Generate a New App Object
 #' 
-#' The Fire generator creates a new 'Fire'-object, which is the the class 
-#' containing all app logic. The class is based on the R6 oo-system and is thus
-#' reference-based with methods and data attached to each object, in contrast to
-#' the more well known S3 and S4 systems.
+#' The Fire generator creates a new `Fire`-object, which is the class containing
+#' all the app logic. The class is based on the [R6][R6::R6Class] OO-system and
+#' is thus reference-based with methods and data attached to each object, in
+#' contrast to the more well known S3 and S4 systems. A `fiery` server is event
+#' driven, which means that it is build up and manipulated by adding event
+#' handlers and triggering events. To learn more about the `fiery` event model,
+#' read the [event documentation][events]. `fiery` servers can be modified 
+#' directly or by attaching plugins. As with events, [plugins has its own
+#' documentation][plugins].
 #' 
 #' @usage NULL
+#' @format NULL
+#' 
+#' @section Initialization:
+#' A new 'Fire'-object is initialized using the `new()` method on the generator:
+#' 
+#' \strong{Usage}
+#' \tabular{l}{
+#'  `app <- Fire$new(host = '127.0.0.1', port = 8080L)`
+#' }
+#' 
+#' \strong{Arguments}
+#' \tabular{lll}{
+#'  `host` \tab  \tab A string overriding the default host (see the *Fields* section below)\cr
+#'  `port` \tab  \tab An integer overriding the default port (see the *Fields* section below)
+#' }
+#' 
+#' *Copying*
+#' 
+#' As `Fire` objects are using reference semantics new copies of an app cannot
+#' be made simply be assigning it to a new variable. If a true copy of a `Fire`
+#' object is desired, use the `clone()` method.
 #' 
 #' @section Fields:
 #' \describe{
-#'  \item{\code{host}}{A string giving a valid IPv4 address owned by the server, or '0.0.0.0' (the default) to listen on all addresses}
-#'  \item{\code{port}}{An integer giving the port number the server should listen on (defaults to 80L)}
-#'  \item{\code{refreshRate}}{The interval in seconds between run cycles when running a blocking server (defaults to 0.001)}
-#'  \item{\code{triggerDir}}{A valid folder where trigger files can be put when running a blocking server (defaults to NULL)}
+#'  \item{`host`}{A string giving a valid IPv4 address owned by the server, or `'0.0.0.0'` to listen on all addresses. The default is `'127.0.0.1'`}
+#'  \item{`port`}{An integer giving the port number the server should listen on (defaults to `8080L`)}
+#'  \item{`refresh_rate`}{The interval in seconds between run cycles when running a blocking server (defaults to `0.001`)}
+#'  \item{`refresh_rate_nb`}{The interval in seconds between run cycles when running a non-bocking server (defaults to `1`)}
+#'  \item{`trigger_dir`}{A valid folder where trigger files can be put when running a blocking server (defaults to `NULL`)}
+#'  \item{`plugins`}{A named list of the already attached plugins. **Static** - can only be modified using the `attach()` method.}
+#'  \item{`root`}{The location of the app. Setting this will remove the root value from requests (or decline them with `400` if the request does not match the root). E.g. the path of a request will be changed from `/demo/test` to `/test` if `root == '/demo'`}
 #' }
 #' 
 #' @section Methods:
 #' \describe{
-#'  \item{\code{ignite(block = TRUE, showcase = FALSE, ...)}}{Begins the server,either blocking the console if \code{block = TRUE} or not. If \code{showcase = TRUE} a browser window is opened directing at the server address. \code{...} will be redirected to the 'start' handler(s)}
-#'  \item{\code{start(block = TRUE, showcase = FALSE, ...)}}{A less dramatic synonym of for \code{ignite}}
-#'  \item{\code{reignite(block = TRUE, showcase = FALSE, ...)}}{As \code{ignite} but additionally triggers the 'resume' event after the 'start' event}
-#'  \item{\code{resume(block = TRUE, showcase = FALSE, ...)}}{Another less dramatic synonym, this time for reignite}
-#'  \item{\code{extinguish()}}{Stops a running server}
-#'  \item{\code{stop()}}{Boring synonym for \code{extinguish}}
-#'  \item{\code{on(event, handler, pos = NULL)}}{Add a handler function to to an event at the given position in the handler stack. Returns a string uniquely identifying the handler}
-#'  \item{\code{off(handlerId)}}{Remove the handler tied to the given id}
-#'  \item{\code{trigger(event, ...)}}{Triggers an event passing the additional arguments to the potential handlers}
-#'  \item{\code{send(message, id)}}{Sends a websocket message to the client with the given id, or to all connected clients if id is missing}
-#'  \item{\code{attach(plugin, ...)}}{Attaches a plugin to the server. A plugin is an R6 object with an \code{onAttach} method}
-#'  \item{\code{header(name, value)}}{Add a global header to the server that will be set on all responses}
-#'  \item{\code{set_data(name, value)}}{Adds data to the servers internal data store}
-#'  \item{\code{get_data(name)}}{Extracts data from the internal data store}
-#'  \item{\code{remove_data(name)}}{Removes the data with the given name from the internal data store}
-#'  \item{\code{time(expr, then, after, loop = FALSE)}}{Add a timed evaluation that will be evaluated after the given number of seconds, potentially repeating if loop=TRUE. After the expression has evaluated the 'then' function will get called with the result of the expression and the server object as arguments.}
-#'  \item{\code{remove_time(id)}}{Removes the timed evaluation identified by the id (returned when adding the evaulation)}
-#'  \item{\code{delay(expr, then)}}{As time except the expr is evaluated immediately at the end of the loop cycle}
-#'  \item{\code{remove_delay(id)}}{Removes the delayed evaluation identified by the id}
-#'  \item{\code{async(expr, then)}}{As delay and time except the expression is evaluated asynchronously. The progress of evaluation is checked at the end of each loop cycle}
-#'  \item{\code{remove_async(id)}}{Removes the async evaluation identified by the id. The evaluation is not necessarily stopped but the then function will not get called.}
-#'  \item{\code{set_client_id_converter(converter)}}{Sets the function that converts an HTTP request into a specific client id}
-#'  \item{\code{test_request(request)}}{Test the result of recieving a specific HTTP request}
-#'  \item{\code{test_header(request)}}{Test the result of recieving a specific HTTP header}
-#'  \item{\code{test_message(request, binary, message, withClose = TRUE)}}{Test the result of recieving a message over websocket and potentially closing the connection afterwards}
-#'  \item{\code{test_websocket(request, message)}}{Test the result of sending a message over websocket}
+#'  \item{`ignite(block = TRUE, showcase = FALSE, ...)`}{Begins the server, either blocking the console if `block = TRUE` or not. If `showcase = TRUE` a browser window is opened directing at the server address. `...` will be redirected to the `start` handler(s)}
+#'  \item{`start(block = TRUE, showcase = FALSE, ...)`}{A less dramatic synonym of for `ignite()`}
+#'  \item{`reignite(block = TRUE, showcase = FALSE, ...)`}{As `ignite` but additionally triggers the `resume` event after the `start` event}
+#'  \item{`resume(block = TRUE, showcase = FALSE, ...)`}{Another less dramatic synonym, this time for `reignite()`}
+#'  \item{`extinguish()`}{Stops a running server}
+#'  \item{`stop()`}{Boring synonym for `extinguish()`}
+#'  \item{`on(event, handler, pos = NULL)`}{Add a `handler` function to to an `event` at the given position (`pos`) in the handler stack. Returns a string uniquely identifying the handler. See the [event documentation][events] for more information.}
+#'  \item{`off(handlerId)`}{Remove the handler tied to the given `id`}
+#'  \item{`trigger(event, ...)`}{Triggers an `event` passing the additional arguments to the potential handlers}
+#'  \item{`send(message, id)`}{Sends a websocket `message` to the client with the given `id`, or to all connected clients if `id` is missing}
+#'  \item{`close_ws_con(id)`}{Closes the websocket connection started from the client with the given `id`, firing the `websocket-closed` event}
+#'  \item{`attach(plugin, ..., force = FALSE)`}{Attaches a `plugin` to the server. See the [plugin documentation][plugins] for more information. Plugins can only get attached once unless `force = TRUE`}
+#'  \item{`has_plugin(name)`}{Check whether a plugin with the given `name` has been attached}
+#'  \item{`header(name, value)`}{Add a global `header` to the server that will be set on all responses. Remove by setting `value = NULL`}
+#'  \item{`set_data(name, value)`}{Adds data to the servers internal data store}
+#'  \item{`get_data(name)`}{Extracts data from the internal data store}
+#'  \item{`remove_data(name)`}{Removes the data with the given `name` from the internal data store}
+#'  \item{`time(expr, then, after, loop = FALSE)`}{Add a timed evaluation (`expr`) that will be evaluated after the given number of seconds (`after`), potentially repeating if `loop = TRUE`. After the expression has evaluated the `then` function will get called with the result of the expression and the server object as arguments.}
+#'  \item{`remove_time(id)`}{Removes the timed evaluation identified by the `id` (returned when adding the evaluation)}
+#'  \item{`delay(expr, then)`}{As `time()` except the `expr` is evaluated immediately at the end of the loop cycle}
+#'  \item{`remove_delay(id)`}{Removes the delayed evaluation identified by the `id`}
+#'  \item{`async(expr, then)`}{As `delay()` and `time()` except the expression is evaluated asynchronously. The progress of evaluation is checked at the end of each loop cycle}
+#'  \item{`remove_async(id)`}{Removes the async evaluation identified by the `id`. The evaluation is not necessarily stopped but the then function will not get called.}
+#'  \item{`set_client_id_converter(converter)`}{Sets the function that converts an HTTP request into a specific client id}
+#'  \item{`clone()`}{Create a copy of the full `Fire` object and return that}
 #' }
 #' 
 #' @importFrom R6 R6Class
-#' @importFrom assertthat is.string is.count is.number has_args assert_that is.dir is.flag has_name
+#' @importFrom assertthat is.string is.count is.number has_args assert_that is.dir is.flag has_name is.error
 #' @importFrom httpuv startServer service startDaemonizedServer stopDaemonizedServer stopServer
 #' @importFrom uuid UUIDgenerate
 #' @importFrom utils browseURL
+#' @importFrom later later
+#' @importFrom stats setNames
+#' @importFrom reqres Request
+#' @importFrom stringi stri_pad_left
 #' 
 #' @export
 #' @docType class
 #' 
+#' @seealso [events] describes how the server event cycle works
+#' 
+#' [plugins] describes how to use plugins to modify the server
+#' 
 #' @examples 
 #' # Create a New App
-#' app <- Fire$new()
-#' app$port <- 4689
+#' app <- Fire$new(port = 4689)
 #' 
-#' # Setup the data everytime it starts
+#' # Setup the data every time it starts
 #' app$on('start', function(server, ...) {
 #'     server$set_data('visits', 0)
 #'     server$set_data('cycles', 0)
@@ -116,7 +151,9 @@ NULL
 Fire <- R6Class('Fire',
     public = list(
         # Methods
-        initialize = function() {
+        initialize = function(host = '127.0.0.1', port = 8080) {
+            self$host <- host
+            self$port <- port
             private$data <- new.env(parent = emptyenv())
             private$handlers <- new.env(parent = emptyenv())
             private$websockets <- new.env(parent = emptyenv())
@@ -125,27 +162,57 @@ Fire <- R6Class('Fire',
             private$TIME <- TimeStack$new()
             private$ASYNC <- AsyncStack$new()
         },
-        ignite = function(block = TRUE, showcase = FALSE, ...) {
+        format = function(...) {
+            text <- c(
+                '\U0001f525 A fiery webserver',
+                '\U0001f525  \U0001f4a5   \U0001f4a5   \U0001f4a5'
+            )
+            mat <- matrix(c('Running on', ': ', paste0(self$host, ':', self$port, self$root)), ncol = 3)
+            plugins <- names(private$pluginList)
+            if (is.null(plugins)) plugins <- 'none'
+            mat <- rbind(mat, c('Plugins attached', ': ', plugins[1]))
+            mat <- rbind(mat, matrix(c(rep('', (length(plugins) - 1)*2), plugins[-1]), ncol = 3))
+            handlers <- lapply(private$handlers, function(x) x$length())
+            if (length(handlers) == 0) {
+                mat <- rbind(mat, c('Event handlers added', ': ', 'none'))
+            } else {
+                mat <- rbind(mat, c('Event handlers added', '', ''))
+                order <- match(names(handlers), private$privateTriggers)
+                order[is.na(order)] <- seq_len(sum(is.na(order))) + max(order, na.rm = TRUE)
+                handlers <- handlers[order(order)]
+                mat <- rbind(mat, matrix(c(names(handlers), rep(': ', length(handlers)), as.character(unlist(handlers))), ncol = 3))
+            }
+            mat[, 1] <- stri_pad_left(mat[, 1], max(nchar(mat[,1])))
+            c(text, paste0('\U0001f525 ', apply(mat, 1, paste, collapse = '')))
+        },
+        ignite = function(block = TRUE, showcase = FALSE, ..., silent = FALSE) {
+            if (!silent) message('Fire started at ', self$host, ':', self$port, self$root)
             private$run(block = block, showcase = showcase, ...)
             invisible(NULL)
         },
-        start = function(block = TRUE, showcase = FALSE, ...) {
-            self$ignite(block = block, showcase = showcase, ...)
+        start = function(block = TRUE, showcase = FALSE, ..., silent = FALSE) {
+            self$ignite(block = block, showcase = showcase, ..., silent = silent)
         },
-        reignite = function(block = TRUE, showcase = FALSE, ...) {
+        reignite = function(block = TRUE, showcase = FALSE, ..., silent = FALSE) {
+            if (!silent) message('Fire restarted at ', self$host, ':', self$port, self$root)
             private$run(block = block, resume = TRUE, showcase = showcase, ...)
             invisible(NULL)
         },
-        resume = function(block = TRUE, showcase = FALSE, ...) {
-            self$reignite(block = block, showcase = showcase, ...)
+        resume = function(block = TRUE, showcase = FALSE, ..., silent = FALSE) {
+            self$reignite(block = block, showcase = showcase, ..., silent = silent)
         },
         extinguish = function() {
             if (private$running) {
                 if (!is.null(private$server)) {
-                    private$running <- FALSE
-                    private$p_trigger('end', server = self)
-                    stopDaemonizedServer(private$server)
-                    private$server <- NULL
+                    if (private$nb_cycle) {
+                        message('Cannot stop server from within a non-blocking event cycle')
+                        flush.console()
+                    } else {
+                        private$running <- FALSE
+                        private$p_trigger('end', server = self)
+                        stopDaemonizedServer(private$server)
+                        private$server <- NULL
+                    }
                 } else {
                     private$quitting <- TRUE
                 }
@@ -182,12 +249,40 @@ Fire <- R6Class('Fire',
         },
         send = function(message, id) {
             private$send_ws(message, id)
-            private$p_trigger('send', server = self, message = message, id = id)
+            private$p_trigger('send', server = self, id = id, message = message)
             invisible(NULL)
         },
-        attach = function(plugin, ...) {
-            plugin$onAttach(self, ...)
+        close_ws_con = function(id) {
+            assert_that(is.string(id))
+            ws <- private$websockets[[id]]
+            if (!is.null(ws)) {
+                private$close_ws(id)
+            }
+        },
+        attach = function(plugin, ..., force = FALSE) {
+            name <- plugin$name
+            assert_that(is.string(name))
+            
+            if (!force && self$has_plugin(name)) {
+                stop('The ', name, ' plugin is already loaded. Use `force = TRUE` to reapply it.', call. = FALSE)
+            }
+            requires <- plugin$require
+            if (!is.null(requires)) {
+                assert_that(is.character(requires))
+                exists <- vapply(requires, self$has_plugin, logical(1))
+                if (!all(exists)) {
+                    stop('The ', name, ' plugin requires the following plugins: ', paste(requires[!exists], collapse = ', '), '.', call. = FALSE)
+                }
+            }
+            has_error <- try(plugin$on_attach(self, ...), silent = TRUE)
+            if (is.error(has_error)) {
+                stop('The ', name, ' plugin failed to attach with the following error: ', has_error, call. = FALSE)
+            }
+            private$add_plugin(plugin, name)
             invisible(NULL)
+        },
+        has_plugin = function(name) {
+            name %in% names(private$pluginList)
         },
         header = function(name, value) {
             assert_that(is.string(name))
@@ -248,17 +343,17 @@ Fire <- R6Class('Fire',
                 close_fun()
             }
         },
-        test_websocket = function(request, message) {
+        test_websocket = function(request, message, close = TRUE) {
             ws <- list(
                 request = request,
                 onMessage = function(func) {},
                 onClose = function(func) {},
                 send = function(message) {message(message)},
-                close = function() {}
+                close = function() {message('closing')}
             )
             private$websocket_logic(ws)
-            self$send(message)
-            private$close_ws(private$client_id(request))
+            self$send(message, private$client_id(request))
+            if (close) private$close_ws(private$client_id(request))
         }
     ),
     active = list(
@@ -272,17 +367,35 @@ Fire <- R6Class('Fire',
             assert_that(is.count(n))
             private$PORT <- n
         },
-        refreshRate = function(rate) {
+        refresh_rate = function(rate) {
             if (missing(rate)) return(private$REFRESHRATE)
             assert_that(is.number(rate))
             private$REFRESHRATE <- rate
         },
-        triggerDir = function(dir) {
+        refresh_rate_nb = function(rate) {
+            if (missing(rate)) return(private$REFRESHRATENB)
+            assert_that(is.number(rate))
+            private$REFRESHRATENB <- rate
+        },
+        trigger_dir = function(dir) {
             if (missing(dir)) return(private$TRIGGERDIR)
             if (!is.null(dir)) {
                 assert_that(is.dir(dir))
             }
             private$TRIGGERDIR <- dir
+        },
+        plugins = function(plugin) {
+            if (!missing(plugin)) {
+                stop('Use the `attach` method to add plugins', call. = FALSE)
+            }
+            private$pluginList
+        },
+        root = function(path) {
+            if (missing(path)) return(private$ROOT)
+            assert_that(is.string(path))
+            path <- sub('/$', '', path)
+            if (path != '') path <- paste0('/', sub('^/+', '', path))
+            private$ROOT <- path
         }
     ),
     private = list(
@@ -290,18 +403,22 @@ Fire <- R6Class('Fire',
         HOST = '127.0.0.1',
         PORT = 8080,
         REFRESHRATE = 0.001,
+        REFRESHRATENB = 1,
         TRIGGERDIR = NULL,
+        ROOT = '',
         
         running = FALSE,
+        nb_cycle = FALSE,
         quitting = FALSE,
-        privateTriggers = c('start', 'resume', 'end', 'cycle-start', 
-                            'cycle-end', 'header', 'before-request', 'request', 
-                            'after-request', 'before-message', 'message', 
-                            'after-message', 'websocket-closed', 'send'),
+        privateTriggers = c('start', 'resume', 'cycle-start', 'header', 
+                            'before-request', 'request', 'after-request', 
+                            'before-message', 'message', 'after-message', 
+                            'websocket-closed', 'send', 'cycle-end', 'end'),
         data = NULL,
         headers = list(),
         handlers = NULL,
         handlerMap = list(),
+        pluginList = list(),
         websockets = NULL,
         server = NULL,
         client_id = NULL,
@@ -320,9 +437,17 @@ Fire <- R6Class('Fire',
             if (!private$running) {
                 private$running <- TRUE
                 private$TIME$reset()
-                private$p_trigger('start', server = self, ...)
+                error <- try(private$p_trigger('start', server = self, ...), silent = TRUE)
+                if (is.error(error)) {
+                    private$running <- FALSE
+                    stop(trimws(error), call. = FALSE)
+                }
                 if (resume) {
-                    private$p_trigger('resume', server = self, ...)
+                    error <- try(private$p_trigger('resume', server = self, ...), silent = TRUE)
+                    if (is.error(error)) {
+                        private$running <- FALSE
+                        stop(trimws(error), call. = FALSE)
+                    }
                 }
                 
                 if (block) {
@@ -367,7 +492,7 @@ Fire <- R6Class('Fire',
                     private$quitting <- FALSE
                     break
                 }
-                Sys.sleep(self$refreshRate)
+                Sys.sleep(self$refresh_rate)
             }
         },
         run_allowing_server = function(showcase = FALSE) {
@@ -384,8 +509,38 @@ Fire <- R6Class('Fire',
             if (showcase) {
                 private$open_browser()
             }
+            
+            private$allowing_cycle()
+        },
+        allowing_cycle = function() {
+            if (private$running) {
+                private$nb_cycle <- TRUE # To hinder stopDeamonizedServer from crashing session
+                private$p_trigger('cycle-start', server = self)
+                private$external_triggers()
+                private$DELAY$eval(server = self)
+                private$TIME$eval(server = self)
+                private$ASYNC$eval(server = self)
+                private$p_trigger('cycle-end', server = self)
+                private$nb_cycle <- FALSE
+                later(function() {
+                    private$allowing_cycle()
+                }, private$REFRESHRATENB)
+            }
+        },
+        mount_request = function(req) {
+            if (!grepl(paste0('^', self$root, '(/|$)'), req$PATH_INFO)) stop('URL not matching mount point', call. = FALSE)
+            req$SCRIPT_NAME <- self$root
+            req$PATH_INFO <- sub(paste0('^', self$root, ''), '', req$PATH_INFO)
+            req
         },
         request_logic = function(req) {
+            request <- try(private$mount_request(req), silent = TRUE)
+            if (is.error(request)) {
+                req <- Request$new(req)
+                return(req$respond()$status_with_text(400L)$as_list())
+            } else {
+                req <- Request$new(request)
+            }
             id <- private$client_id(req)
             args <- unlist(
                 unname(
@@ -394,40 +549,45 @@ Fire <- R6Class('Fire',
                 ), 
                 recursive = FALSE
             )
-            if (is.null(args)) {
-                args <- list(
-                    event = 'request',
-                    server = self,
-                    id = id,
-                    request = req
-                )
-            } else {
-                args <- modifyList(args, list(
-                    event = 'request',
-                    server = self,
-                    id = id,
-                    request = req
-                ))
-            }
-            response <- tail(do.call(private$p_trigger, args), 1)[[1]]
-            if (is.null(response)) response <- notFound
-            if (!(is.list(response) && all(has_name(response, c('status', 'headers', 'body'))))) {
-                response <- serverError
-            }
-            response$headers <- modifyList(private$headers, response$headers)
-            private$p_trigger('after-request', server = self, id = id, request = req, response = response)
+            private$p_trigger('request', server = self, id = id, request = req, arg_list = args)
+            response <- req$respond()
+            for (i in names(private$headers)) response$set_header(i, private$headers[[i]])
+            response <- response$as_list()
+            private$p_trigger('after-request', server = self, id = id, request = req)
             response
         },
         header_logic = function(req) {
+            request <- try(private$mount_request(req), silent = TRUE)
+            if (is.error(request)) {
+                req <- Request$new(req)
+                return(req$respond()$status_with_text(400L)$as_list())
+            } else {
+                req <- Request$new(request)
+            }
             id <- private$client_id(req)
-            private$p_trigger('header', server = self, id = id, request = req)
+            response <- private$p_trigger('header', server = self, id = id, request = req)
+            if (length(response) == 0) {
+                NULL
+            } else {
+                continue <- tail(response, 1)[[1]]
+                assert_that(is.flag(continue))
+                if (continue) NULL
+                else req$respond()$as_list()
+            }
         },
         websocket_logic = function(ws) {
-            id <- private$client_id(ws$request)
+            request <- try(private$mount_request(ws$request), silent = TRUE)
+            if (is.error(request)) {
+                ws$close()
+                return()
+            } else {
+                req <- Request$new(request)
+            }
+            id <- private$client_id(req)
             assign(id, ws, envir = private$websockets)
             
-            ws$onMessage(private$message_logic(id, ws$request))
-            ws$onClose(private$close_ws_logic(id, ws$request))
+            ws$onMessage(private$message_logic(id, req))
+            ws$onClose(private$close_ws_logic(id, req))
         },
         message_logic = function(id, request) {
             function(binary, msg) {
@@ -439,16 +599,14 @@ Fire <- R6Class('Fire',
                     ),
                     recursive = FALSE
                 )
-                args <- modifyList(list(binary = binary, message = msg), args)
-                args <- modifyList(args, list(
-                    event = 'message',
-                    server = self,
-                    id = id,
-                    request = request
-                ))
-                do.call(private$p_trigger, args)
+                if (is.null(args)) args <- structure(list(), names = character())
+                if ('binary' %in% names(args)) binary <- args$binary
+                if ('message' %in% names(args)) msg <- args$message
+                args <- modifyList(args, list(binary = NULL, message = NULL))
                 
-                private$p_trigger('after-message', server = self, id = id, binary = args$binary, message = args$message, request = request)
+                private$p_trigger('message', server = self, id = id, binary = binary, message = msg, request = request, arg_list = args)
+                
+                private$p_trigger('after-message', server = self, id = id, binary = binary, message = msg, request = request)
             }
         },
         close_ws_logic = function(id, request) {
@@ -466,11 +624,14 @@ Fire <- R6Class('Fire',
             event <- private$handlerMap[[id]]
             private$handlers[[event]]$remove(id)
         },
+        add_plugin = function(plugin, name) {
+            private$pluginList[[name]] <- plugin
+        },
         p_trigger = function(event, ...) {
             if (!is.null(private$handlers[[event]])) {
                 private$handlers[[event]]$dispatch(...)
             } else {
-                `names<-`(list(), character())
+                setNames(list(), character())
             }
         },
         external_triggers = function() {
@@ -504,12 +665,15 @@ Fire <- R6Class('Fire',
                 id <- ls(envir = private$websockets)
             }
             for (i in id) {
-                private$websockets[[i]]$send(message)
+                if (i %in% names(private$websockets)) private$websockets[[i]]$send(message)
             }
         },
         close_ws = function(id) {
-            private$websockets[[id]]$close()
-            rm(list = id, envir = private$websockets)
+            ws <- private$websockets[[id]]
+            if (!is.null(ws)) {
+                try(ws$close(), silent = TRUE)
+                rm(list = id, envir = private$websockets)
+            }
         },
         open_browser = function() {
             url <- paste0('http://', private$HOST, ':', private$PORT, '/')