+++ tests/testthat/test-Fire.R
@@ -22,6 +22,26 @@ test_that('handlers can be added, triggered and removed', {
     expect_named(triggerRes, character())
 })
 
+test_that('Fire objects are printed', {
+    app <- Fire$new()
+    app$attach(list(
+        name = 'test',
+        on_attach = function(...) {}
+    ))
+    app$on('start', function(...){})
+    app$on('request', function(...){})
+    app$on('request', function(...){})
+    
+    expect_output(print(app), 'A fiery webserver')
+    expect_output(print(app), 'Plugins attached: test')
+    expect_output(print(app), 'start: 1')
+    expect_output(print(app), 'request: 2')
+    
+    app <- Fire$new()
+    expect_output(print(app), 'Plugins attached: none')
+    expect_output(print(app), 'Event handlers added: none')
+})
+
 test_that('protected events cannot be triggered', {
     app <- Fire$new()
     
@@ -51,12 +71,30 @@ test_that('plugins are being attached', {
     app <- Fire$new()
     app$set_data('test', 10)
     plugin <- list(
-        onAttach = function(server, extraPar) {
+        on_attach = function(server, extraPar) {
             server$on('test', function(...){10 + extraPar})
         }
     )
+    expect_error(app$attach(plugin, 15))
+    plugin$name <- 'plugin'
     app$attach(plugin, 15)
+    expect_error(app$attach(plugin, 10))
     expect_equal(app$trigger('test')[[1]], 25)
+    plugin2 <- list(
+        on_attach = function(...) {message('test')},
+        name = 'plugin2',
+        require = c('plugin', 'test')
+    )
+    expect_error(app$attach(plugin2))
+    plugin2$require <- 'plugin'
+    expect_message(app$attach(plugin2), 'test')
+    expect_equal(plugin, app$plugins$plugin)
+    expect_error(app$plugins$test <- plugin)
+    plugin3 <- list(
+        on_attach = function(...) {stop('test')},
+        name = 'plugin3'
+    )
+    expect_error(app$attach(plugin3), 'The plugin3 plugin failed to attach with the following error:')
 })
 
 test_that('id converter can be set and gets called', {
@@ -66,7 +104,7 @@ test_that('id converter can be set and gets called', {
     })
     request <- fake_request('http://www.example.com', REMOTE_ADDR = '127.0.0.1')
     app$test_request(request)
-    expect_equal(app$get_data('id'), client_to_id(request))
+    expect_equal(app$get_data('id'), client_to_id(reqres::Request$new(request)))
     
     app$set_client_id_converter(function(request) {
         10
@@ -90,19 +128,24 @@ test_that('active bindings work', {
     app$port <- 10
     expect_equal(app$port, 10)
     
-    expect_error(app$refreshRate <- 'test')
-    expect_error(app$refreshRate <- 1:5)
-    app$refreshRate <- 10.5
-    expect_equal(app$refreshRate, 10.5)
+    expect_error(app$refresh_rate <- 'test')
+    expect_error(app$refresh_rate <- 1:5)
+    app$refresh_rate <- 10.5
+    expect_equal(app$refresh_rate, 10.5)
     
-    expect_error(app$triggerDir <- 'test')
-    expect_error(app$triggerDir <- 1:5)
+    expect_error(app$refresh_rate_nb <- 'test')
+    expect_error(app$refresh_rate_nb <- 1:5)
+    app$refresh_rate_nb <- 10.5
+    expect_equal(app$refresh_rate_nb, 10.5)
+    
+    expect_error(app$trigger_dir <- 'test')
+    expect_error(app$trigger_dir <- 1:5)
     dir <- tempdir()
-    app$triggerDir <- dir
-    expect_equal(app$triggerDir, dir)
+    app$trigger_dir <- dir
+    expect_equal(app$trigger_dir, dir)
 })
 
-test_that('lifecycle events get fired', {
+test_that('life cycle events get fired', {
     app <- Fire$new()
     app$on('start', function(server, ...) {
         server$set_data('events', c(server$get_data('events'), 'start'))
@@ -141,15 +184,17 @@ test_that('lifecycle events get fired', {
     expect_equal(reigniteRes, c('start', 'resume', 'cycle-start', 'cycle-end', 'cycle-start', 'cycle-end', 'end'))
     
     app$ignite(block = FALSE)
+    Sys.sleep(2)
+    expect_message(later::run_now(), 'Cannot stop server from within a non-blocking event cycle')
     app$stop()
     igniteResNoBlock <- app$get_data('events')
     app$remove_data('events')
-    expect_equal(igniteResNoBlock, c('start', 'end'))
+    expect_equal(igniteResNoBlock, c('start', 'cycle-start', 'cycle-end', 'cycle-start', 'cycle-end', 'end'))
     app$reignite(block = FALSE)
     app$extinguish()
     reigniteResNoBlock <- app$get_data('events')
     app$remove_data('events')
-    expect_equal(reigniteResNoBlock, c('start', 'resume', 'end'))
+    expect_equal(reigniteResNoBlock, c('start', 'resume', 'cycle-start', 'cycle-end', 'end'))
 })
 
 test_that('request events fire', {
@@ -160,14 +205,14 @@ test_that('request events fire', {
         server$set_data('events', c(server$get_data('events'), 'before'))
         list(test = 4)
     })
-    app$on('request', function(server, ...) {
+    app$on('request', function(server, arg_list, id, request, ...) {
         server$set_data('events', c(server$get_data('events'), 'during'))
-        server$set_data('passed_args', list(...)$test)
-        list(status = 200)
+        server$set_data('passed_args', arg_list$test)
+        request$respond()$status_with_text(200L)
     })
-    app$on('after-request', function(server, response, ...) {
+    app$on('after-request', function(server, request, ...) {
         server$set_data('events', c(server$get_data('events'), 'after'))
-        server$set_data('passed_response', response)
+        server$set_data('passed_response', request$respond()$as_list())
     })
     response <- app$test_request(request)
     
@@ -182,11 +227,11 @@ test_that('message events fire', {
     
     app$on('before-message', function(server, ...) {
         server$set_data('events', c(server$get_data('events'), 'before'))
-        list(test = 4, message = 'test2')
+        list(test = 4, message = 'test2', binary = FALSE)
     })
-    app$on('message', function(server, ...) {
+    app$on('message', function(server, message, arg_list, ...) {
         server$set_data('events', c(server$get_data('events'), 'during'))
-        server$set_data('passed_args', list(...)[c('test', 'message')])
+        server$set_data('passed_args', list(test = arg_list$test, message = message))
     })
     app$on('after-message', function(server, response, ...) {
         server$set_data('events', c(server$get_data('events'), 'after'))
@@ -195,17 +240,51 @@ test_that('message events fire', {
     
     expect_equal(app$get_data('events'), c('before', 'during', 'after'))
     expect_equal(app$get_data('passed_args'), list(test = 4, message = 'test2'))
+    
+    app <- Fire$new()
+    request <- fake_request('http://www.example.com')
+    
+    app$on('message', function(server, message, arg_list, ...) {
+        server$set_data('events', c(server$get_data('events'), 'during'))
+        server$set_data('passed_args', list(test = arg_list$test, message = message))
+    })
+    app$on('after-message', function(server, response, ...) {
+        server$set_data('events', c(server$get_data('events'), 'after'))
+    })
+    app$test_message(request, FALSE, 'test')
+    
+    expect_equal(app$get_data('events'), c('during', 'after'))
 })
 
 test_that('header event fire', {
     app <- Fire$new()
     request <- fake_request('http://www.example.com')
-    
+    expect_true(is.null(app$test_header(request)))
     app$on('header', function(server, ...) {
         server$set_data('header', TRUE)
+        TRUE
     })
     app$test_header(request)
     expect_true(app$get_data('header'))
+    app$on('header', function(server, request, ...) {
+        request$respond()$status_with_text(500L)
+        FALSE
+    })
+    response <- app$test_header(request)
+    expect_equal(response$status, 500L)
+})
+
+test_that('errors in start and resume gets caught', {
+    app <- Fire$new()
+    app$on('start', function(...) {
+        stop('Testing an error')
+    })
+    expect_error(app$ignite(silent = TRUE))
+    app <- Fire$new()
+    app$on('resume', function(...) {
+        stop('Testing an error')
+    })
+    expect_error(app$reignite(silent = TRUE))
 })
 
 test_that('futures can be added and called', {
@@ -232,7 +311,7 @@ test_that('futures can be added and called', {
         server$extinguish()
     })
     app$remove_delay(id)
-    expect_silent(app$ignite())
+    expect_silent(app$ignite(silent = TRUE))
 
     app <- Fire$new()
     app$time({
@@ -241,18 +320,15 @@ test_that('futures can be added and called', {
         message(res)
         server$extinguish()
     }, 1)
-    app$on('start', function(server, ...) server$set_data('count', 0))
+    app$on('start', function(server, ...) server$set_data('time', Sys.time()))
     app$on('cycle-end', function(server, ...) {
-        cycle <- server$get_data('count')
-        if (cycle > 100) {
+        start <- server$get_data('time')
+        if (Sys.time() - start > 2) {
             server$extinguish()
-        } else {
-            server$set_data('count', cycle + 1)
         }
     })
     expect_message(app$ignite(), '10')
-    
-    skip_on_os('windows')
+
     app <- Fire$new()
     id <- app$time({
         10
@@ -261,18 +337,16 @@ test_that('futures can be added and called', {
         server$extinguish()
     }, 1)
     app$remove_time(id)
-    app$on('start', function(server, ...) server$set_data('count', 0))
+    app$on('start', function(server, ...) server$set_data('time', Sys.time()))
     app$on('cycle-end', function(server, ...) {
-        cycle <- server$get_data('count')
-        if (cycle > 100) {
+        start <- server$get_data('time')
+        if (Sys.time() - start > 2) {
             server$extinguish()
-        } else {
-            server$set_data('count', cycle + 1)
         }
     })
-    expect_silent(app$ignite())
+    expect_silent(app$ignite(silent = TRUE))
 
-    # The async stuff fail on windows builders though it works fine locally
+    skip_on_os('windows') # The async stuff fail on windows builders though it works fine locally
     app <- Fire$new()
     id <- app$async({
         10
@@ -281,16 +355,14 @@ test_that('futures can be added and called', {
         server$extinguish()
     })
     app$remove_async(id)
-    app$on('start', function(server, ...) server$set_data('count', 0))
+    app$on('start', function(server, ...) server$set_data('time', Sys.time()))
     app$on('cycle-end', function(server, ...) {
-        cycle <- server$get_data('count')
-        if (cycle > 100) {
+        start <- server$get_data('time')
+        if (Sys.time() - start > 2) {
             server$extinguish()
-        } else {
-            server$set_data('count', cycle + 1)
         }
     })
-    expect_silent(app$ignite())
+    expect_silent(app$ignite(silent = TRUE))
     
     app <- Fire$new()
     app$async({
@@ -299,13 +371,11 @@ test_that('futures can be added and called', {
         message(res)
         server$extinguish()
     })
-    app$on('start', function(server, ...) server$set_data('count', 0))
+    app$on('start', function(server, ...) server$set_data('time', Sys.time()))
     app$on('cycle-end', function(server, ...) {
-        cycle <- server$get_data('count')
-        if (cycle > 1000) {
+        start <- server$get_data('time')
+        if (Sys.time() - start > 2) {
             server$extinguish()
-        } else {
-            server$set_data('count', cycle + 1)
         }
     })
     expect_message(app$ignite(), '10')
@@ -323,7 +393,7 @@ test_that('external triggers are fired', {
     app <- Fire$new()
     
     dir <- tempdir()
-    app$triggerDir <- dir
+    app$trigger_dir <- dir
     
     app$on('test', function(server, ...) {
         server$set_data('ext_args', list(...))
@@ -337,11 +407,17 @@ test_that('external triggers are fired', {
 
 test_that('websockets are attached, and removed', {
     app <- Fire$new()
-    
+    req <- fake_request('http://www.example.com')
     app$on('send', function(server, ...) {server$set_data('send', TRUE)})
     expect_null(app$get_data('send'))
-    expect_message(app$test_websocket(fake_request('http://www.example.com'), 'test'), 'test')
+    expect_message(app$test_websocket(req, 'test', FALSE), 'test')
     expect_true(app$get_data('send'))
+    expect_message(app$send('keep testing', client_to_id(req)), 'keep testing')
+    expect_message(app$send('keep testing again'), 'keep testing again')
+    expect_message(app$close_ws_con(client_to_id(req)), 'closing')
+    expect_silent(app$close_ws_con(client_to_id(req)))
+    expect_silent(app$send('keep testing', client_to_id(req)))
+    expect_silent(app$send('keep testing again'))
 })
 
 test_that('showcase opens a browser', {
@@ -360,17 +436,40 @@ test_that('global headers are assigned and used', {
     app <- Fire$new()
     app$header('X-Powered-By', 'fiery')
     app$header('X-XSS-Protection', '1; mode=block')
-    app$on('request', function(...) {
-        list(
-            status = 200L,
-            headers = list(),
-            body = 'test'
-        )
+    app$on('request', function(request, ...) {
+        request$respond()$status_with_text(200L)
     })
     response <- app$test_request(fake_request('www.example.com'))
-    expect_equal(response$headers, list('X-Powered-By' = 'fiery', 'X-XSS-Protection' = '1; mode=block'))
+    expect_equal(response$headers, list('Content-Type' = 'text/plain', 'X-Powered-By' = 'fiery', 'X-XSS-Protection' = '1; mode=block'))
     app$header('X-XSS-Protection', NULL)
     response <- app$test_request(fake_request('www.example.com'))
-    expect_equal(response$headers, list('X-Powered-By' = 'fiery'))
+    expect_equal(response$headers, list('Content-Type' = 'text/plain', 'X-Powered-By' = 'fiery'))
     expect_equal(app$header('X-Powered-By'), 'fiery')
+})
+
+test_that('app can be mounted at path', {
+    app <- Fire$new()
+    expect_equal(app$root, '')
+    expect_error(app$root <- 123)
+    expect_error(app$root <- c('test', 'test2'))
+    app$root <- '//test/'
+    expect_equal(app$root, '/test')
+    req <- fake_request('http://example.com/test/testing')
+    app$test_request(req)
+    expect_equal(req$PATH_INFO, '/testing')
+    req <- fake_request('http://example.com/test/testing')
+    app$test_header(req)
+    expect_equal(req$PATH_INFO, '/testing')
+    req <- fake_request('http://example.com/test/testing')
+    expect_message(app$test_websocket(req, 'test'), 'test')
+    expect_equal(req$PATH_INFO, '/testing')
+    
+    req <- fake_request('http://example.com/testing')
+    res <- app$test_request(req)
+    expect_equal(res$status, 400L)
+    req <- fake_request('http://example.com/testing')
+    res <- app$test_header(req)
+    expect_equal(res$status, 400L)
+    req <- fake_request('http://example.com/testing')
+    expect_message(app$test_websocket(req, 'test'), '^closing\n$')
 })